pyproject.toml
[project]
name = "sql-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.4.0",
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "openai>=2.21.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
